pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "syntho_hive"
version = "0.1.0"
description = "Production Synthetic Data Engine with Relational Integrity"
authors = [
  { name = "SynthoHive Team", email = "team@synthohive.dev" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "torch>=1.10.0",
    "pyspark>=3.2.0",
    "pydantic>=2.0.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "faker>=13.0.0",
    "delta-spark>=2.0.0",
    "structlog>=21.1.0",
    "pyarrow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "flake8",
    "mypy"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
